refactor: move UI classes to the ui package
